pull conversion logic into Features directory
Also move tests closer to their directories
