+++ src/NUnitToXUnit/Properties/AssemblyInfo.cs
@@ -2,7 +2,6 @@
 // Licensed under the MIT license. See the LICENSE.md file in the project root for license information.
 
 using System.Reflection;
-using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 
 // General Information about an assembly is controlled through the following
