+++ src/NUnitToXUnit/Visitor/MemberAccessVisitor.cs
@@ -1,12 +1,10 @@
 ﻿// Copyright (c) 2018 Jetabroad Pty Limited. All Rights Reserved.
 // Licensed under the MIT license. See the LICENSE.md file in the project root for license information.
 
-using System.Collections.Generic;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
 using Microsoft.CodeAnalysis.CSharp.Syntax;
-using NUnitToXUnit.Extensions;
-using static Microsoft.CodeAnalysis.CSharp.SyntaxFactory;
+using NUnitToXUnit.Features.ConvertAsserts;
 
 namespace NUnitToXUnit.Visitor
 {
@@ -14,39 +12,9 @@ namespace NUnitToXUnit.Visitor
     {
         public override SyntaxNode VisitMemberAccessExpression(MemberAccessExpressionSyntax node)
         {
+            var convertedAsserts = ConvertAssertStatements.Convert(node);
             requires.XUnit = true;
-            return base.VisitMemberAccessExpression(ReplaceMemberAccessExpression(node));
+            return base.VisitMemberAccessExpression(convertedAsserts);
         }
-
-        private static MemberAccessExpressionSyntax ReplaceMemberAccessExpression(MemberAccessExpressionSyntax node)
-        {
-            return node.IsAssertExpression() && Assertions.ContainsKey(node.Name.ToString())
-                ? CreateAssertExpression(node)
-                : node;
-        }
-
-        private static MemberAccessExpressionSyntax CreateAssertExpression(MemberAccessExpressionSyntax node)
-        {
-            return MemberAccessExpression(
-                    SyntaxKind.SimpleMemberAccessExpression,
-                    IdentifierName("Assert").WithTriviaFrom(node),
-                    IdentifierName(Assertions[node.Name.ToString()]));
-        }
-
-        private static readonly IReadOnlyDictionary<string, string> Assertions = new Dictionary<string, string>
-        {
-            ["AreEqual"] = "Equal",
-            ["AreNotEqual"] = "NotEqual",
-            ["AreNotSame"] = "NotSame",
-            ["AreSame"] = "Same",
-            ["IsEmpty"] = "Empty",
-            ["IsFalse"] = "False",
-            ["IsInstanceOf"] = "IsType",
-            ["IsNotEmpty"] = "NotEmpty",
-            ["IsNotInstanceOf"] = "IsNotType",
-            ["IsNotNull"] = "NotNull",
-            ["IsNull"] = "Null",
-            ["IsTrue"] = "True"
-        };
     }
 }
